+++ pappl/printer-ipp.c
@@ -931,7 +931,9 @@ _papplPrinterSetAttributes(
   ipp_attribute_t	*rattr;		// Current request attribute
   ipp_tag_t		value_tag;	// Value tag
   cups_len_t		count;		// Number of values
-  const char		*name;		// Attribute name
+  const char		*name,		// Attribute name
+			*keyword;	// Keyword value
+  int			intvalue;	// Integer value
   char			defname[128],	// xxx-default name
 			value[1024];	// xxx-default value
   cups_len_t		i, j;		// Looping vars
@@ -1047,21 +1049,67 @@ _papplPrinterSetAttributes(
     // Then copy the xxx-default values to the
     if (!strcmp(name, "identify-actions-default"))
     {
-      driver_data.identify_default = PAPPL_IDENTIFY_ACTIONS_NONE;
+      pappl_identify_actions_t	identify_actions = PAPPL_IDENTIFY_ACTIONS_NONE;
+					// "identify-actions" bit values
 
       for (i = 0, count = ippGetCount(rattr); i < count; i ++)
-        driver_data.identify_default |= _papplIdentifyActionsValue(ippGetString(rattr, i, NULL));
-      do_defaults = true;
+      {
+        pappl_identify_actions_t action;// Current action
+
+        keyword = ippGetString(rattr, i, NULL);
+        action  = _papplIdentifyActionsValue(keyword);
+
+        if (!action || !(action & driver_data.identify_supported))
+        {
+	  papplClientRespondIPP(client, IPP_STATUS_ERROR_ATTRIBUTES_OR_VALUES, "Unsupported \"identify-actions-default\" value '%s'.", keyword);
+	  break;
+        }
+
+        identify_actions |= action;
+      }
+
+      if (i < count)
+      {
+        papplClientRespondIPPUnsupported(client, rattr);
+      }
+      else
+      {
+	driver_data.identify_default = identify_actions;
+	do_defaults                  = true;
+      }
     }
     else if (!strcmp(name, "label-mode-configured"))
     {
-      driver_data.mode_configured = _papplLabelModeValue(ippGetString(rattr, 0, NULL));
-      do_defaults = true;
+      pappl_label_mode_t label_mode;	// "label-mode-configured" value
+
+      keyword    = ippGetString(rattr, 0, NULL);
+      label_mode = _papplLabelModeValue(keyword);
+
+      if (!(label_mode & driver_data.mode_supported))
+      {
+	papplClientRespondIPP(client, IPP_STATUS_ERROR_ATTRIBUTES_OR_VALUES, "Unsupported \"label-mode-configured\" value '%s'.", keyword);
+        papplClientRespondIPPUnsupported(client, rattr);
+      }
+      else
+      {
+	driver_data.mode_configured = label_mode;
+	do_defaults                 = true;
+      }
     }
     else if (!strcmp(name, "label-tear-offset-configured"))
     {
-      driver_data.tear_offset_configured = ippGetInteger(rattr, 0);
-      do_defaults = true;
+      intvalue = ippGetInteger(rattr, 0);
+
+      if (intvalue < driver_data.tear_offset_supported[0] || intvalue > driver_data.tear_offset_supported[1])
+      {
+	papplClientRespondIPP(client, IPP_STATUS_ERROR_ATTRIBUTES_OR_VALUES, "Unsupported \"label-tear-offset-configured\" value '%d'.", intvalue);
+        papplClientRespondIPPUnsupported(client, rattr);
+      }
+      else
+      {
+	driver_data.tear_offset_configured = intvalue;
+	do_defaults                        = true;
+      }
     }
     else if (!strcmp(name, "media-col-default"))
     {
@@ -1121,8 +1169,7 @@ _papplPrinterSetAttributes(
     }
     else if (!strcmp(name, "output-bin-default"))
     {
-      const char *keyword = ippGetString(rattr, 0, NULL);
-					// Keyword value
+      keyword = ippGetString(rattr, 0, NULL);
 
       for (i = 0; i < driver_data.num_bin; i ++)
       {
@@ -1140,33 +1187,102 @@ _papplPrinterSetAttributes(
     }
     else if (!strcmp(name, "print-color-mode-default"))
     {
-      driver_data.color_default = _papplColorModeValue(ippGetString(rattr, 0, NULL));
-      do_defaults = true;
+      pappl_color_mode_t color_mode;	// "print-color-mode" bit value
+
+      keyword    = ippGetString(rattr, 0, NULL);
+      color_mode = _papplColorModeValue(keyword);
+
+      if (!(color_mode & driver_data.color_supported))
+      {
+        papplClientRespondIPP(client, IPP_STATUS_ERROR_ATTRIBUTES_OR_VALUES, "Unsupported \"print-color-mode-default\" value '%s'.", keyword);
+        papplClientRespondIPPUnsupported(client, rattr);
+      }
+      else
+      {
+	driver_data.color_default = color_mode;
+	do_defaults               = true;
+      }
     }
     else if (!strcmp(name, "print-content-optimize-default"))
     {
-      driver_data.content_default = _papplContentValue(ippGetString(rattr, 0, NULL));
-      do_defaults = true;
+      pappl_content_t content;		// "print-content-optimize" bit value
+
+      keyword = ippGetString(rattr, 0, NULL);
+      content = _papplContentValue(keyword);
+
+      if (!content)
+      {
+        papplClientRespondIPP(client, IPP_STATUS_ERROR_ATTRIBUTES_OR_VALUES, "Unsupported \"print-content-optimize-default\" value '%s'.", keyword);
+        papplClientRespondIPPUnsupported(client, rattr);
+      }
+      else
+      {
+	driver_data.content_default = content;
+	do_defaults                 = true;
+      }
     }
     else if (!strcmp(name, "print-darkness-default"))
     {
-      driver_data.darkness_default = ippGetInteger(rattr, 0);
-      do_defaults = true;
+      intvalue = ippGetInteger(rattr, 0);
+
+      if (intvalue < 0 || intvalue > 100 || !driver_data.darkness_supported)
+      {
+        papplClientRespondIPP(client, IPP_STATUS_ERROR_ATTRIBUTES_OR_VALUES, "Unsupported \"print-darkness-default\" value '%d'.", intvalue);
+        papplClientRespondIPPUnsupported(client, rattr);
+      }
+      else
+      {
+	driver_data.darkness_default = intvalue;
+	do_defaults                  = true;
+      }
     }
     else if (!strcmp(name, "print-quality-default"))
     {
-      driver_data.quality_default = (ipp_quality_t)ippGetInteger(rattr, 0);
-      do_defaults = true;
+      intvalue = ippGetInteger(rattr, 0);
+
+      if (intvalue < IPP_QUALITY_DRAFT || intvalue > IPP_QUALITY_HIGH)
+      {
+        papplClientRespondIPP(client, IPP_STATUS_ERROR_ATTRIBUTES_OR_VALUES, "Unsupported \"print-quality-default\" value '%d'.", intvalue);
+        papplClientRespondIPPUnsupported(client, rattr);
+      }
+      else
+      {
+	driver_data.quality_default = (ipp_quality_t)intvalue;
+	do_defaults                 = true;
+      }
     }
     else if (!strcmp(name, "print-scaling-default"))
     {
-      driver_data.scaling_default = _papplScalingValue(ippGetString(rattr, 0, NULL));
-      do_defaults = true;
+      pappl_scaling_t scaling;		// "print-scaling" bit value
+
+      keyword = ippGetString(rattr, 0, NULL);
+      scaling = _papplScalingValue(keyword);
+
+      if (!scaling)
+      {
+        papplClientRespondIPP(client, IPP_STATUS_ERROR_ATTRIBUTES_OR_VALUES, "Unsupported \"print-scaling-default\" value '%s'.", keyword);
+        papplClientRespondIPPUnsupported(client, rattr);
+      }
+      else
+      {
+        driver_data.scaling_default = scaling;
+        do_defaults                 = true;
+      }
     }
     else if (!strcmp(name, "print-speed-default"))
     {
-      driver_data.speed_default = ippGetInteger(rattr, 0);
-      do_defaults = true;
+      intvalue = ippGetInteger(rattr, 0);
+
+      if (intvalue < driver_data.speed_supported[0] || intvalue > driver_data.speed_supported[1])
+      {
+        papplClientRespondIPP(client, IPP_STATUS_ERROR_ATTRIBUTES_OR_VALUES, "Unsupported \"print-speed-default\" value '%d'.", intvalue);
+        papplClientRespondIPPUnsupported(client, rattr);
+      }
+      else
+      {
+	driver_data.speed_default = intvalue;
+	do_defaults               = true;
+      }
     }
     else if (!strcmp(name, "printer-contact-col"))
     {
@@ -1175,8 +1291,18 @@ _papplPrinterSetAttributes(
     }
     else if (!strcmp(name, "printer-darkness-configured"))
     {
-      driver_data.darkness_configured = ippGetInteger(rattr, 0);
-      do_defaults = true;
+      intvalue = ippGetInteger(rattr, 0);
+
+      if (intvalue < 0 || intvalue > 100 || !driver_data.darkness_supported)
+      {
+        papplClientRespondIPP(client, IPP_STATUS_ERROR_ATTRIBUTES_OR_VALUES, "Unsupported \"printer-darkness-configured\" value '%d'.", intvalue);
+        papplClientRespondIPPUnsupported(client, rattr);
+      }
+      else
+      {
+	driver_data.darkness_configured = intvalue;
+	do_defaults                     = true;
+      }
     }
     else if (!strcmp(name, "printer-geo-location"))
     {
@@ -1200,10 +1326,28 @@ _papplPrinterSetAttributes(
     }
     else if (!strcmp(name, "printer-resolution-default"))
     {
+      int	xres, yres;		// X and Y resolution
       ipp_res_t units;			// Resolution units
 
-      driver_data.x_default = ippGetResolution(rattr, 0, &driver_data.y_default, &units);
-      do_defaults = true;
+      xres = ippGetResolution(rattr, 0, &yres, &units);
+
+      for (i = 0; i < driver_data.num_resolution; i ++)
+      {
+        if (xres == driver_data.x_resolution[i] && yres == driver_data.y_resolution[i])
+          break;
+      }
+
+      if (units != IPP_RES_PER_INCH || i >= driver_data.num_resolution)
+      {
+        papplClientRespondIPP(client, IPP_STATUS_ERROR_ATTRIBUTES_OR_VALUES, "Unsupported \"printer-resolution-default\" value.");
+        papplClientRespondIPPUnsupported(client, rattr);
+      }
+      else
+      {
+	driver_data.x_default = xres;
+	driver_data.y_default = yres;
+	do_defaults           = true;
+      }
     }
     else if (!strcmp(name, "printer-wifi-password"))
     {
@@ -1229,16 +1373,19 @@ _papplPrinterSetAttributes(
     }
     else if (!strcmp(name, "sides-default"))
     {
-      pappl_sides_t sides_default = _papplSidesValue(ippGetString(rattr, 0, NULL));
-					// Sides value
+      pappl_sides_t sides;		// Sides value
 
-      if (!sides_default || !(driver_data.sides_supported & sides_default))
+      keyword = ippGetString(rattr, 0, NULL);
+      sides   = _papplSidesValue(keyword);
+
+      if (!sides || !(driver_data.sides_supported & sides))
       {
-        papplClientRespondIPP(client, IPP_STATUS_ERROR_ATTRIBUTES_OR_VALUES, "Unsupported \"sides-default\" value '%s'.", ippGetString(rattr, 0, NULL));
+        papplClientRespondIPP(client, IPP_STATUS_ERROR_ATTRIBUTES_OR_VALUES, "Unsupported \"sides-default\" value '%s'.", keyword);
+	papplClientRespondIPPUnsupported(client, rattr);
       }
       else
       {
-        driver_data.sides_default = sides_default;
+        driver_data.sides_default = sides;
         do_defaults               = true;
       }
     }